+++ Model/Ingenico/RequestBuilder/Common/OrderBuilder.php
@@ -5,15 +5,13 @@ namespace Ingenico\Connect\Model\Ingenico\RequestBuilder\Common;
 use Ingenico\Connect\Helper\Data as DataHelper;
 use Ingenico\Connect\Model\ConfigInterface;
 use Ingenico\Connect\Model\Ingenico\MerchantReference;
+use Ingenico\Connect\Model\Ingenico\RequestBuilder\Common\Order\AdditionalInputBuilder;
 use Ingenico\Connect\Model\Ingenico\RequestBuilder\Common\Order\CustomerBuilder;
 use Ingenico\Connect\Model\Ingenico\RequestBuilder\Common\Order\ShippingBuilder;
 use Ingenico\Connect\Model\Ingenico\RequestBuilder\Common\Order\ShoppingCartBuilder;
 use Ingenico\Connect\Sdk\Domain\Definitions\AmountOfMoneyFactory;
-use Ingenico\Connect\Sdk\Domain\Payment\Definitions\AdditionalOrderInputFactory;
 use Ingenico\Connect\Sdk\Domain\Payment\Definitions\OrderFactory;
 use Ingenico\Connect\Sdk\Domain\Payment\Definitions\OrderReferencesFactory;
-use Ingenico\Connect\Sdk\Domain\Payment\Definitions\OrderTypeInformationFactory;
-use Magento\Framework\Stdlib\DateTime\DateTime;
 use Magento\Sales\Model\Order;
 
 /**
@@ -37,11 +35,6 @@ class OrderBuilder
     private $shoppingCartBuilder;
 
     /**
-     * @var DateTime
-     */
-    private $dateTime;
-
-    /**
      * @var OrderFactory
      */
     private $orderFactory;
@@ -57,16 +50,6 @@ class OrderBuilder
     private $orderReferencesFactory;
 
     /**
-     * @var AdditionalOrderInputFactory
-     */
-    private $additionalOrderInputFactory;
-
-    /**
-     * @var OrderTypeInformationFactory
-     */
-    private $orderTypeInformationFactory;
-
-    /**
      * @var MerchantReference
      */
     private $merchantReference;
@@ -75,6 +58,10 @@ class OrderBuilder
      * @var ShippingBuilder
      */
     private $shippingBuilder;
+    /**
+     * @var AdditionalInputBuilder
+     */
+    private $additionalInputBuilder;
 
     /**
      * OrderBuilder constructor.
@@ -82,38 +69,33 @@ class OrderBuilder
      * @param ConfigInterface $ePaymentsConfig
      * @param CustomerBuilder $customerBuilder
      * @param ShoppingCartBuilder $shoppingCartBuilder
-     * @param DateTime $dateTime
      * @param OrderFactory $orderFactory
      * @param AmountOfMoneyFactory $amountOfMoneyFactory
      * @param OrderReferencesFactory $orderReferencesFactory
-     * @param AdditionalOrderInputFactory $additionalOrderInputFactory
-     * @param OrderTypeInformationFactory $orderTypeInformationFactory
+     * @param AdditionalInputBuilder $additionalInputBuilder
+     * @param ShippingBuilder $shippingBuilder
      * @param MerchantReference $merchantReference
      */
     public function __construct(
         ConfigInterface $ePaymentsConfig,
         CustomerBuilder $customerBuilder,
         ShoppingCartBuilder $shoppingCartBuilder,
-        DateTime $dateTime,
         OrderFactory $orderFactory,
         AmountOfMoneyFactory $amountOfMoneyFactory,
         OrderReferencesFactory $orderReferencesFactory,
-        AdditionalOrderInputFactory $additionalOrderInputFactory,
-        OrderTypeInformationFactory $orderTypeInformationFactory,
+        AdditionalInputBuilder $additionalInputBuilder,
         ShippingBuilder $shippingBuilder,
         MerchantReference $merchantReference
     ) {
         $this->ePaymentsConfig = $ePaymentsConfig;
         $this->customerBuilder = $customerBuilder;
         $this->shoppingCartBuilder = $shoppingCartBuilder;
-        $this->dateTime = $dateTime;
         $this->orderFactory = $orderFactory;
         $this->amountOfMoneyFactory = $amountOfMoneyFactory;
         $this->orderReferencesFactory = $orderReferencesFactory;
-        $this->additionalOrderInputFactory = $additionalOrderInputFactory;
-        $this->orderTypeInformationFactory = $orderTypeInformationFactory;
         $this->merchantReference = $merchantReference;
         $this->shippingBuilder = $shippingBuilder;
+        $this->additionalInputBuilder = $additionalInputBuilder;
     }
 
     /**
@@ -129,7 +111,7 @@ class OrderBuilder
 
         $ingenicoOrder->shoppingCart = $this->shoppingCartBuilder->create($order);
         $ingenicoOrder->references = $this->getReferences($order);
-        $ingenicoOrder->additionalInput = $this->getAdditionalInput($order);
+        $ingenicoOrder->additionalInput = $this->additionalInputBuilder->create($order);
         $ingenicoOrder->shipping = $this->shippingBuilder->create($order);
 
         return $ingenicoOrder;
@@ -159,23 +141,5 @@ class OrderBuilder
         $references->descriptor = $this->ePaymentsConfig->getDescriptor();
 
         return $references;
-    }
-
-    /**
-     * @param Order $order
-     * @return \Ingenico\Connect\Sdk\Domain\Payment\Definitions\AdditionalOrderInput
-     */
-    private function getAdditionalInput(Order $order)
-    {
-        $additionalInput = $this->additionalOrderInputFactory->create();
-
-        $additionalInput->orderDate = $this->dateTime->date('YmdHis', strtotime($order->getCreatedAt()));
-
-        $typeInformation = $this->orderTypeInformationFactory->create();
-        $typeInformation->purchaseType = 'good';
-        $typeInformation->usageType = 'commercial';
-        $additionalInput->typeInformation = $typeInformation;
-
-        return $additionalInput;
     }
 }